slc: Fix file descriptor leak
Arthur Lapz reported that he is seeing a blocking socket from the
public client side after the app side has closed the connection.

This turned out that we haven't closed the SLC socket file descriptor:

```
     Isolated                              Server
     --------                              ------

      fd_app --                         -- fd_public_client
               |----[ SLC circuit ]----|
  fd_circuit --                         -- fd_circuit

```
After the app in the isolated client has closed its fd, we didn't close
the fd_app and fd_circuit. That's why the server is still sleeping,
waiting for any response from the isolated area.

This was just a typo in if statements that caused the close() paths were
not taken. Simply fix the typo. Change "==" to "!=".

Link: https://t.me/lostcontrol_s1/16737
Link: https://t.me/lostcontrol_s1/16739 # Reproducer
Fixes: 04a86adffa415e8365d27929bbb826028e00bfd3 ("Initial commit")
Reported-by: Arthur Lapz <rlapz@gnuweeb.org>
Reviewed-by: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>
Link: https://lore.gnuweeb.org/gwml/20220606022747.316158-1-ammarfaizi2%40gnuweeb.org
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>